requirements.txt
attrs==22.1.0
aws-cdk-lib==2.48.0
boto3==1.25.5
botocore==1.28.5
cattrs==22.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
constructs==10.1.141
exceptiongroup==1.0.0
idna==3.4
jmespath==1.0.1
jsii==1.70.0
publication==0.0.3
python-dateutil==2.8.2
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
typeguard==2.13.3
typing_extensions==4.4.0
urllib3==1.26.12
